chore: Remove go & shared directories/remove old ollama trash(#270)
